refactor: 拆分Stimulus契约测试职责
